+++ NEWS.md
@@ -1,5 +1,11 @@
 # greta.gp (development version)
 
+# greta.gp 0.2.2
+
+* Add "rlang" to package imports.
+* update internals to use `cli_abort`
+* depend on greta version 0.5.0 and R version 0.4.1
+
 # greta.gp 0.2.1
 
 * Use sentinel "_PACKAGE"